Return Nil rather than '' from .get at EOF
I noticed when using an IO::Blob object with Text::CSV.getline that it
wasn't ending (returning an empty list,) when it reached EOF.

It appears that Text::CSV is testing for the definedness of the return
of .get, which is consistent with the documented behaviour of the core
IO::Handle.get.

So this adjusts .get to return Nil at EOF rather than the empty string.